+++ blockchain/src/blocks/mod.rs
@@ -1,10 +1,8 @@
 mod block;
-mod block_msg;
 mod errors;
 mod ticket;
 mod tipset;
 
 pub use block::*;
-pub use block_msg::*;
 pub use errors::*;
 pub use tipset::*;
