+++ blockchain/src/chain_sync/mod.rs
@@ -1,1 +1,4 @@
+mod block_msg;
 mod block_proposer;
+
+pub use block_msg::*;
